requirements.txt
discord.py>=2.0.0
aiohttp>=3.8.0
pyyaml>=6.0
